Fixed bug in loop sumarisation and houdini verifier.
Summary: Fixed a bug that made loop sumarisation ignore loop assertions. Still
need to check if the assertions are in the middle of the loop and fail
in this case.

Remove a bug in Houdini Verifier that made is loop forever if the failed
invariants discovered where in another loop than the one examined.

We still need to fix a bug in predication for havoc in scope.

Test Plan: Tests pass

Reviewers: me